+++ crates/scouter_server/src/api/mod.rs
@@ -1,4 +1,3 @@
-pub mod handler;
 pub mod middleware;
 pub mod router;
 pub mod setup;
