+++ crates/scouter_server/src/api/router.rs
@@ -1,7 +1,7 @@
 
 use crate::api::middleware::track_metrics;
 use crate::api::state::AppState;
-use crate::api::routes::{get_health_router, get_drift_router, get_profile_router, get_alert_router};
+use crate::api::routes::{get_health_router, get_drift_router, get_profile_router, get_alert_router, get_observability_router};
 use anyhow::Result;
 use axum::http::{
     header::{ACCEPT, AUTHORIZATION, CONTENT_TYPE},
@@ -15,6 +15,17 @@ use tower_http::cors::CorsLayer;
 
 const ROUTE_PREFIX: &str = "/scouter";
 
+/// Create the main router for the application
+/// 
+/// This function creates the main router for the application by merging all the sub-routers
+/// and adding the necessary middleware.
+/// 
+/// # Parameters
+/// - `app_state` - The application state shared across all handlers
+/// 
+/// # Returns
+/// 
+/// The main router for the application
 pub async fn create_router(app_state: Arc<AppState>) -> Result<Router> {
     let cors = CorsLayer::new()
         .allow_methods([Method::GET, Method::PUT, Method::DELETE])
@@ -25,41 +36,19 @@ pub async fn create_router(app_state: Arc<AppState>) -> Result<Router> {
         let drift_routes = get_drift_router(ROUTE_PREFIX).await?;
         let profile_routes = get_profile_router(ROUTE_PREFIX).await?;
         let alert_routes = get_alert_router(ROUTE_PREFIX).await?;
+        let observability_routes = get_observability_router(ROUTE_PREFIX).await?;
 
         let merged_routes = Router::new()
         .merge(health_routes)
         .merge(drift_routes)
         .merge(profile_routes)
         .merge(alert_routes)
+        .merge(observability_routes)
         .route_layer(middleware::from_fn(track_metrics));
 
         Ok(Router::new()
         .merge(merged_routes)
         .layer(cors)
         .with_state(app_state))
-    // let router = Router::new()
-    //outer::new()
-    //   .route(&format!("{}/healthcheck", ROUTE_PREFIX), get(health_check))
-    //   .route(
-    //       &format!("{}/drift", ROUTE_PREFIX),
-    //       get(get_drift).post(insert_drift),
-    //   )
-    //   .route(
-    //       &format!("{}/profile", ROUTE_PREFIX),
-    //       post(insert_drift_profile)
-    //           .put(update_drift_profile)
-    //           .get(get_profile),
-    //   )
-    //   .route(
-    //       &format!("{}/profile/status", ROUTE_PREFIX),
-    //       put(update_drift_profile_status),
-    //   )
-    //   .route(&format!("{}/alerts", ROUTE_PREFIX), get(get_drift_alerts))
-    //   .route(
-    //       &format!("{}/observability/metrics", ROUTE_PREFIX),
-    //       get(get_observability_metrics),
-    //   )
-    //   .route_layer(middleware::from_fn(track_metrics))
-    //   .with_state(app_state)
-    //   .layer(cors)
+    
 }
