+++ crates/scouter_server/src/api/routes/health/mod.rs
@@ -1,5 +1,4 @@
-pub mod schema;
+
 pub mod route;
 
-pub use schema::*;
 pub use route::*;
