+++ crates/scouter_server/src/api/routes/mod.rs
@@ -4,8 +4,8 @@ pub mod profile;
 pub mod alerts;
 pub mod observability;
 
-pub use health::*;
-pub use drift::*;
-pub use profile::*;
-pub use alerts::*;
-pub use observability::*;
+pub use health::get_health_router;
+pub use drift::get_drift_router;
+pub use profile::get_profile_router;
+pub use alerts::get_alert_router;
+pub use observability::get_observability_router;
