+++ crates/scouter_server/src/api/routes/observability/mod.rs
@@ -1,0 +1,3 @@
+pub mod route;
+
+pub use route::*;
