+++ crates/scouter_server/src/api/setup.rs
@@ -7,6 +7,9 @@ use tracing_subscriber::fmt::time::UtcTime;
 const DEFAULT_TIME_PATTERN: &str =
     "[year]-[month]-[day]T[hour repr:24]:[minute]:[second]::[subsecond digits:4]";
 
+/// Setup logging for the application
+/// 
+/// This function initializes the logging system for the application
 pub async fn setup_logging() -> Result<(), anyhow::Error> {
     let time_format = time::format_description::parse(DEFAULT_TIME_PATTERN).unwrap();
