+++ crates/scouter_server/src/main.rs
@@ -18,6 +18,7 @@ use crate::consumer::kafka::startup::kafka_startup::startup_kafka;
 #[cfg(feature = "rabbitmq")]
 use crate::consumer::rabbitmq::startup::rabbitmq_startup::startup_rabbitmq;
 
+/// Start the metrics server for prometheus
 async fn start_metrics_server() -> Result<(), anyhow::Error> {
     let app = metrics_app().with_context(|| "Failed to setup metrics app")?;
 
@@ -32,6 +33,8 @@ async fn start_metrics_server() -> Result<(), anyhow::Error> {
     Ok(())
 }
 
+
+/// Start the main server
 async fn start_main_server() -> Result<(), anyhow::Error> {
     // setup logging
     setup_logging()
@@ -102,50 +105,50 @@ async fn main() -> Result<(), anyhow::Error> {
     Ok(())
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::{
-        body::Body,
-        http::{Request, StatusCode},
-    };
-    use http_body_util::BodyExt;
-    use serde_json::Value;
-    use tower::ServiceExt; // for `call`, `oneshot`, and `ready`
-
-    #[tokio::test]
-    async fn test_health_check() {
-        let pool = create_db_pool(Some(
-            "postgresql://postgres:admin@localhost:5432/scouter?".to_string(),
-        ))
-        .await
-        .with_context(|| "Failed to create Postgres client")
-        .unwrap();
-
-        let db_client = sql::postgres::PostgresClient::new(pool).unwrap();
-
-        let app = create_router(Arc::new(AppState {
-            db: db_client.clone(),
-        }));
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .uri("/scouter/healthcheck")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        //assert response
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-
-        let v: Value = serde_json::from_str(std::str::from_utf8(&body[..]).unwrap()).unwrap();
-
-        let message: &str = v.get("message").unwrap().as_str().unwrap();
-
-        assert_eq!(message, "Alive");
-    }
-}
+//#[cfg(test)]
+//mod tests {
+//    use super::*;
+//    use axum::{
+//        body::Body,
+//        http::{Request, StatusCode},
+//    };
+//    use http_body_util::BodyExt;
+//    use serde_json::Value;
+//    use tower::ServiceExt; // for `call`, `oneshot`, and `ready`
+//
+//    #[tokio::test]
+//    async fn test_health_check() {
+//        let pool = create_db_pool(Some(
+//            "postgresql://postgres:admin@localhost:5432/scouter?".to_string(),
+//        ))
+//        .await
+//        .with_context(|| "Failed to create Postgres client")
+//        .unwrap();
+//
+//        let db_client = sql::postgres::PostgresClient::new(pool).unwrap();
+//
+//        let app = create_router(Arc::new(AppState {
+//            db: db_client.clone(),
+//        }));
+//
+//        let response = app
+//            .oneshot(
+//                Request::builder()
+//                    .uri("/scouter/healthcheck")
+//                    .body(Body::empty())
+//                    .unwrap(),
+//            )
+//            .await
+//            .unwrap();
+//
+//        //assert response
+//        assert_eq!(response.status(), StatusCode::OK);
+//        let body = response.into_body().collect().await.unwrap().to_bytes();
+//
+//        let v: Value = serde_json::from_str(std::str::from_utf8(&body[..]).unwrap()).unwrap();
+//
+//        let message: &str = v.get("message").unwrap().as_str().unwrap();
+//
+//        assert_eq!(message, "Alive");
+//    }
+//}
